MainViewModel과 DeprecatedMainViewModel을 스위치
이전에 pagination을 하던 DeprecatedMainViewModel을 MainViewModel로, on-device AI를 사용하던 MainViewModel을 DeprecatedMainViewModel로 바꿨다.
