+++ MemoWithTags/MemoWithTags/Presentation/ViewModel/DeprecatedMainViewModel.swift
@@ -1,168 +1,197 @@
 //
-//  DeprecatedMainViewModel.swift
+//  MainViewModel.swift
 //  MemoWithTags
 //
-//  Created by Swimming Ryu on 1/30/25.
+//  Created by Swimming Ryu on 1/5/25.
 //
 
+// 이전에 on-device AI와 fileManager를 사용하고 pagination을 사용하지 않던 코드
 
-// 이전애 서버에서 memo와 tag를 받아와서 pagination하던 코드
 /*
 import Foundation
 import SwiftUI
+import Accelerate
 
 @MainActor
 final class MainViewModel: BaseViewModel, ObservableObject {
     
     @Published var isLoading: Bool = false
     
-    //mainPage 변수들
+    // 여기 있는 모든 리스트들은 정렬되지 않은 리스트다.
+    // 단, recommendingMemos, recommendingTags, searchedMemos, searchedTag는 "관련성 기준"으로 정렬되어있다.
+    
+    // MARK: - Main Page Variables
     @Published var memos: [Memo] = []
     @Published var tags: [Tag] = []
-    @Published var mainCurrentPage: Int = 0
-    @Published var mainTotalPages: Int = 1
+    @Published var recommendingMemos: [Memo] = []
+    @Published var recommendingTags: [Tag] = []
+    // 참고: 사용자가 태그 검색 창에 검색을 한다고 recommendingTags는 변하지 않는다.
+    // tagRecommendation에 recommendingTags가 그대로 보이는 것이 아니라, editor에 있는 tag를 제외하고 검색어와 안 맞는 tag를 제외하고 Create Tag를 하는 등 추가적인 과정이 있다.
     
-    //searchPage 변수들
+    // MARK: - Search Page Variables
     @Published var searchBarText: String = ""
     @Published var searchBarSelectedTags: [Tag] = []
     @Published var searchedMemos: [Memo] = []
     @Published var searchedTags: [Tag] = []
-    @Published var searchCurrentPage: Int = 0
-    @Published var searchTotalPages: Int = 1
     
-    // editor의 변수들 (축소, 확대 상태 모두)
+    // MARK: - Editor Variables (Both Collapsed and Expanded States)
     @Published var editorState: EditorState = .create
     @Published var editorContent: String = ""
+    @Published var editorTagSearchBarText: String = ""
     @Published var editorTags: [Tag] = []
-    enum EditorState {
+    
+    enum EditorState: Equatable {
         case create
         case update(target: Memo)
     }
     
-    // 메모 정렬과 관련된 변수
+    // MARK: - Memo Sorting Variables
     @Published var sortMemo: Sort = .byCreate
-    @Published var sortSearch: Sort = .byCreate
+    
     enum Sort {
         case byCreate
         case byUpdate
     }
     
-    // mainView에서 첫 검색을 할 때나, pagenation을 할 때 모두 사용된다.
-    func fetchMemos() async {
-        guard !isLoading else { return }
-        
-        isLoading = true
-        mainCurrentPage += 1
-        
-        guard mainCurrentPage <= mainTotalPages else {
-            isLoading = false
-            mainCurrentPage -= 1
-            return
-        }
-        
-        let result = await useCases.fetchMemoUseCase.execute(content: nil, tagIds: nil, dateRange: nil, page: mainCurrentPage)
+    // MARK: - Memo Scroll Target in MainView
+    
+    /// scrollTaget이 0이면 아무 것도 강조되지 않는다.
+    /// 1부터 recommendedTags.count까지 있으면 그것 안에서 scroll이 되면서 강조된다.
+    /// AI Recommend에서 사용한다.
+    @Published var scrollTarget: Int = 0
+    /// 기본적으로는 nil이다.
+    /// searchView에서 넘어올 때 사용한다.
+    @Published var scrollSingleTarget: UUID? = nil
+    
+    // MARK: - Load and Save Opeartions between Filesystem
+    
+    func loadMemosAndTagsFromFileSystem() async {
+        let result = await useCases.loadMemosAndTagsUseCase.execute()
         
         switch result {
-        case .success(let paginatedMemos):
-            let updatedMemos = paginatedMemos.memos.map { memo -> Memo in
-                var updatedMemo = memo
-                updatedMemo.tags = getTags(from: updatedMemo.tagIds)
-                return updatedMemo
-            }
-            
-            self.memos.append(contentsOf: updatedMemos)
-            self.mainTotalPages = paginatedMemos.totalPages
-            
+        case .success(let data):
+            print("Successfully loaded Memos and Tags.")
+            self.memos = data.memos
+            self.tags = data.tags
         case .failure(let error):
-            appState.system.showAlert = true
-            appState.system.errorMessage = error.localizedDescription()
+            switch error {
+            case .fileNotFound:
+                // 파일이 존재하지 않는 경우, 빈 배열로 초기화하고 파일 생성
+                print("Files not found. Initializing memos.json and tags.json.")
+                self.memos = []
+                self.tags = []
+                await saveMemosAndTagsToFileSystem()
+            default:
+                appState.system.showAlert = true
+                appState.system.errorMessage = error.errorDescription ?? error.localizedDescription
+            }
         }
-        
-        isLoading = false
     }
     
-    // searchView에서 첫 검색을 할 때나, pagenation을 할 때 모두 사용된다.
-    func searchMemos(content: String? = nil, tagIds: [Int]? = nil, dateRange: ClosedRange<Date>? = nil) async {
-        guard !isLoading else { return }
-        
-        isLoading = true
-        searchCurrentPage += 1
-        
-        guard searchCurrentPage <= searchTotalPages else {
-            isLoading = false
-            searchCurrentPage -= 1
-            return
-        }
-        
-        let result = await useCases.fetchMemoUseCase.execute(content: content, tagIds: tagIds, dateRange: dateRange, page: self.searchCurrentPage)
+    func saveMemosAndTagsToFileSystem() async {
+        let result = await useCases.saveMemosAndTagsUseCase.execute(memos: self.memos, tags: self.tags)
         
         switch result {
-        case .success(let paginatedMemos):
-            let updatedMemos = paginatedMemos.memos.map { memo -> Memo in
-                var updatedMemo = memo
-                updatedMemo.tags = getTags(from: updatedMemo.tagIds)
-                return updatedMemo
-            }
-            
-            self.searchedMemos.append(contentsOf: updatedMemos)
-            self.searchTotalPages = paginatedMemos.totalPages
-
+        case .success():
+            print("Memos and Tags successfully saved to filesystem.")
         case .failure(let error):
-            // SearchView에서 wait이 끝나고 searchMemo가 실행되는 와중에 새로운 Task가 생성되어서 Task가 사라지면 MemoError.unknown이 뜬다. 이것은 정상적인 결과이기 때문에 무시한다.
-            if (error != MemoError.unknown) {
+            print("Failed to save Memos and Tags.")
+            appState.system.showAlert = true
+            appState.system.errorMessage = error.errorDescription ?? error.localizedDescription
+        }
+    }
+    
+    // MARK: - CRUD Operations for Memos
+    
+    /// 메모를 생성할 때 임베딩 벡터를 생성하고, 관련 태그의 임베딩 벡터를 업데이트합니다.
+    func createMemo(content: String, tagIds: [UUID], locked: Bool) async {
+        isLoading = true
+        
+        do {
+            let newId = UUID()
+            
+            let embeddingVector = try await createEmbeddingVectorWithAI(text: content)
+            
+            let currentDate = Date()
+            
+            let result = await useCases.createMemoUseCase.execute(
+                id: newId,
+                content: content,
+                tagIds: tagIds,
+                locked: locked,
+                embeddingVector: embeddingVector,
+                createdAt: currentDate,
+                updatedAt: currentDate
+            )
+            
+            switch result {
+            case .success(let memo):
+                self.memos.append(memo)
+                
+                // 관련 태그의 임베딩 벡터 업데이트
+                await updateAssociatedTagsEmbedding(for: memo)
+                
+                await saveMemosAndTagsToFileSystem()
+                
+            case .failure(let error):
                 appState.system.showAlert = true
                 appState.system.errorMessage = error.localizedDescription()
             }
-        }
-        
-        isLoading = false
-    }
-    
-    func createMemo(content: String, tagIds: [Int], locked: Bool) async {
-        isLoading = true
-        
-        let result = await useCases.createMemoUseCase.execute(content: content, tagIds: tagIds, locked: locked)
-        
-        switch result {
-        case .success(let memo):
-            var memoWithFilledTags = memo
-            memoWithFilledTags.tags = getTags(from: memoWithFilledTags.tagIds)
-            self.memos.insert(memoWithFilledTags, at: 0)
-        case .failure(let error):
+        } catch {
             appState.system.showAlert = true
-            appState.system.errorMessage = error.localizedDescription()
+            appState.system.errorMessage = "메모 생성 중 오류가 발생했습니다: \(error.localizedDescription)"
         }
         
         isLoading = false
     }
     
-    func updateMemo(memoId: Int, content: String, tagIds: [Int], locked: Bool) async {
+    /// 메모를 업데이트할 때 임베딩 벡터를 생성하고, 관련 태그의 임베딩 벡터를 업데이트합니다.
+    func updateMemo(id: UUID, content: String, tagIds: [UUID], locked: Bool) async {
         isLoading = true
         
-        let result = await useCases.updateMemoUseCase.execute(memoId: memoId, content: content, tagIds: tagIds, locked: locked)
-        switch result {
-        case .success(let memo):
-            var memoWithFilledTags = memo
-            memoWithFilledTags.tags = getTags(from: memoWithFilledTags.tagIds)
-            if let index = self.memos.firstIndex(where: { $0.id == memoId }) {
-                self.memos[index] = memoWithFilledTags
+        do {
+            let updatingMemoIndex = self.memos.firstIndex(where: { $0.id == id })!
+            
+            let embeddingVector = try await createEmbeddingVectorWithAI(text: content)
+            
+            let currentDate = Date()
+            
+            let result = await useCases.updateMemoUseCase.execute(
+                id: id,
+                content: content,
+                tagIds: tagIds,
+                locked: locked,
+                embeddingVector: embeddingVector,
+                createdAt: self.memos[updatingMemoIndex].createdAt,
+                updatedAt: currentDate
+            )
+            
+            switch result {
+            case .success(let memo):
+                self.memos[updatingMemoIndex] = memo
+                await saveMemosAndTagsToFileSystem()
+            case .failure(let error):
+                appState.system.showAlert = true
+                appState.system.errorMessage = error.localizedDescription()
             }
-        case .failure(let error):
+        } catch {
             appState.system.showAlert = true
-            appState.system.errorMessage = error.localizedDescription()
+            appState.system.errorMessage = "메모 업데이트 중 오류가 발생했습니다: \(error.localizedDescription)"
         }
         
         isLoading = false
     }
     
-    func deleteMemo(memoId: Int) async {
+    /// 메모를 삭제합니다.
+    func deleteMemo(id: UUID) async {
         isLoading = true
         
-        let result = await useCases.deleteMemoUseCase.execute(memoId: memoId)
+        let result = await useCases.deleteMemoUseCase.execute(id: id)
         switch result {
         case .success:
-            self.memos.removeAll { $0.id == memoId }
-            self.searchedMemos.removeAll { $0.id == memoId }
+            self.memos.removeAll { $0.id == id }
+            self.searchedMemos.removeAll { $0.id == id }
+            await saveMemosAndTagsToFileSystem()
         case .failure(let error):
             appState.system.showAlert = true
             appState.system.errorMessage = error.localizedDescription()
@@ -171,88 +200,107 @@ final class MainViewModel: BaseViewModel, ObservableObject {
         isLoading = false
     }
     
-    func fetchTags() async {
-        isLoading = true
-        
-        let result = await useCases.fetchTagUseCase.execute()
-        switch result {
-        case .success(let fetchedTags):
-            self.tags = fetchedTags
-        case .failure(let error):
-            appState.system.showAlert = true
-            appState.system.errorMessage = error.localizedDescription()
-        }
-        
-        isLoading = false
-    }
+    // MARK: - CRUD Operations for Tags
     
+    /// 태그를 생성할 때 임베딩 벡터를 생성합니다.
     func createTag(name: String, color: Color.TagColor) async {
         isLoading = true
         
-        let result = await useCases.createTagUseCase.execute(name: name, color: color)
-        switch result {
-        case .success(let tag):
-            self.tags.append(tag)
-//            self.editingMemoSelectedTags.append(tag)
-        case .failure(let error):
-            appState.system.showAlert = true
-            appState.system.errorMessage = error.localizedDescription()
-        }
-        
-        isLoading = false
-    }
-    
-    func updateTag(tagId: Int, name: String, color: Color.TagColor) async {
-        isLoading = true
-        
-        let result = await useCases.updateTagUseCase.execute(tagId: tagId, name: name, color: color)
-        switch result {
-        case .success(let tag):
-            // Main과 Search의 tag 변경
-            if let index = self.tags.firstIndex(where: { $0.id == tagId }) {
-                self.tags[index] = tag
-            }
-            if let index = self.searchedTags.firstIndex(where: { $0.id == tagId }) {
-                self.searchedTags[index] = tag
-            }
+        do {
+            let newId = UUID()
             
-            // Main과 Search의 tag 변경
-            for index in memos.indices {
-                if let tagIndex = memos[index].tags.firstIndex(where: { $0.id == tagId }) {
-                    memos[index].tags[tagIndex] = tag
-                }
+            let embeddingVector = try await createEmbeddingVectorWithAI(text: name)
+            
+            let currentDate = Date()
+            
+            let result = await useCases.createTagUseCase.execute(
+                id: newId,
+                name: name,
+                color: color,
+                embeddingVector: embeddingVector,
+                createdAt: currentDate,
+                updatedAt: currentDate
+            )
+            
+            switch result {
+            case .success(let tag):
+                self.tags.append(tag)
+                self.editorTags.append(tag)
+                // 현재 수정하고 있는 메모에 tag를 추가해야 한다.
+                await saveMemosAndTagsToFileSystem()
+                
+                recommendMemosAndTags()
+            case .failure(let error):
+                appState.system.showAlert = true
+                appState.system.errorMessage = error.localizedDescription()
             }
-            // Main과 Search의 memo에 있는 tag 변경
-            for index in searchedMemos.indices {
-                if let tagIndex = searchedMemos[index].tags.firstIndex(where: { $0.id == tagId }) {
-                    searchedMemos[index].tags[tagIndex] = tag
-                }
-            }
-        case .failure(let error):
+        } catch {
             appState.system.showAlert = true
-            appState.system.errorMessage = error.localizedDescription()
+            appState.system.errorMessage = "태그 생성 중 오류가 발생했습니다: \(error.localizedDescription)"
         }
         
         isLoading = false
     }
     
-    func deleteTag(tagId: Int) async {
+    /// 태그를 업데이트할 때 임베딩 벡터를 생성합니다.
+    func updateTag(id: UUID, name: String, color: Color.TagColor) async {
         isLoading = true
         
-        let result = await useCases.deleteTagUseCase.execute(tagId: tagId)
+        do {
+            let updatingTagIndex = self.tags.firstIndex(where: { $0.id == id })!
+            
+            let embeddingVector = try await createEmbeddingVectorWithAI(text: name)
+            
+            let currentDate = Date()
+            
+            let result = await useCases.updateTagUseCase.execute(
+                id: id,
+                name: name,
+                color: color,
+                embeddingVector: embeddingVector,
+                createdAt: self.tags[updatingTagIndex].createdAt,
+                updatedAt: currentDate
+            )
+            
+            switch result {
+            case .success(let tag):
+                self.tags[updatingTagIndex] = tag
+                await saveMemosAndTagsToFileSystem()
+                
+                recommendMemosAndTags()
+            case .failure(let error):
+                appState.system.showAlert = true
+                appState.system.errorMessage = error.localizedDescription()
+            }
+        } catch {
+            appState.system.showAlert = true
+            appState.system.errorMessage = "태그 업데이트 중 오류가 발생했습니다: \(error.localizedDescription)"
+        }
+        
+        isLoading = false
+    }
+    
+    /// 태그를 삭제합니다.
+    func deleteTag(id: UUID) async {
+        isLoading = true
+        
+        let result = await useCases.deleteTagUseCase.execute(id: id)
         switch result {
         case .success:
-            // Main과 Search의 tag 삭제
-            self.tags.removeAll { $0.id == tagId }
-            self.searchedTags.removeAll { $0.id == tagId }
+            // Main과 Search의 태그 삭제
+            self.tags.removeAll { $0.id == id }
+            self.searchedTags.removeAll { $0.id == id }
             
-            // Main과 Search의 memo에 있는 tag 삭제
+            // Main과 Search의 메모에서 해당 태그 삭제
             for index in memos.indices {
-                self.memos[index].tags.removeAll { $0.id == tagId }
+                self.memos[index].tagIds.removeAll { $0 == id }
             }
             for index in searchedMemos.indices {
-                self.searchedMemos[index].tags.removeAll { $0.id == tagId }
+                self.searchedMemos[index].tagIds.removeAll { $0 == id }
             }
+            await saveMemosAndTagsToFileSystem()
+            
+            recommendMemosAndTags()
         case .failure(let error):
             appState.system.showAlert = true
             appState.system.errorMessage = error.localizedDescription()
@@ -261,17 +309,19 @@ final class MainViewModel: BaseViewModel, ObservableObject {
         isLoading = false
     }
     
-    ///main view에서 onApear때 쓰는 함수
-    func initMemo() async {
-        if tags.isEmpty {
-            await fetchTags()
-        }
-        if memos.isEmpty {
-            await fetchMemos()
+    // MARK: - Initialization and User Info
+    
+    /// Main View가 나타날 때 호출되는 초기화 함수
+    func initMainViewModel() async {
+        await getUserInfo()
+        guard let userId = appState.user.userId else { return }
+        useCases.userChangedUseCase.execute(userId: userId)
+        if memos.isEmpty || tags.isEmpty {
+            await loadMemosAndTagsFromFileSystem()
         }
     }
     
-    ///settings view에서 유저정보 가져오는 함수
+    /// Settings View에서 유저 정보를 가져오는 함수
     func getUserInfo() async {
         isLoading = true
         
@@ -280,8 +330,11 @@ final class MainViewModel: BaseViewModel, ObservableObject {
         switch result {
         case .success(let user):
             appState.user.userId = user.id
+            appState.user.userNumber = user.userNumber
             appState.user.userName = user.nickname
             appState.user.userEmail = user.email
+            appState.user.isSocial = user.isSocial
+            
         case .failure(let error):
             appState.system.showAlert = true
             appState.system.errorMessage = error.localizedDescription()
@@ -290,7 +343,7 @@ final class MainViewModel: BaseViewModel, ObservableObject {
         isLoading = false
     }
     
-    ///settings view에서 로그아웃하는 함수
+    /// Settings View에서 로그아웃하는 함수
     func logout() async {
         let result = await useCases.logoutUseCase.execute()
         
@@ -303,72 +356,343 @@ final class MainViewModel: BaseViewModel, ObservableObject {
             appState.user.userId = nil
             appState.user.userName = nil
             appState.user.userEmail = nil
+            appState.user.isSocial = false
             
             appState.navigation.reset()
             appState.navigation.push(to: .root)
+            
         case .failure(let error):
             appState.system.showAlert = true
             appState.system.errorMessage = error.localizedDescription()
         }
     }
     
-    ///태그 추천 해주는 함수: editor에 들어간 것들 뺴고
-    func recommendTags() -> [Tag] {
-        tags.filter { !editorTags.contains($0) }
-    }
-    
-    ///editor에서 submit 했을 때 작동
+    /// Editor에서 Submit 했을 때 작동
     func submit() async {
+        hideKeyboard()
+        
         let trimmedContent = editorContent.trimmingCharacters(in: .whitespacesAndNewlines)
         guard !trimmedContent.isEmpty else { return }
-    
         let tagIds = editorTags.map { $0.id }
         
         switch editorState {
         case .create:
             await createMemo(content: trimmedContent, tagIds: tagIds, locked: false)
-            memos = []
-            mainCurrentPage = 0
-            await fetchMemos()
 
         case .update(let target):
-            await updateMemo(memoId: target.id, content: trimmedContent, tagIds: tagIds, locked: target.locked)
+            await updateMemo(id: target.id, content: trimmedContent, tagIds: tagIds, locked: target.locked)
         }
         
-        // Reset the input fields
+        // 입력 필드 초기화
         editorState = .create
         editorContent = ""
         editorTags = []
-        hideKeyboard()
     }
     
+    func withdrawal(email: String) async {
+        let result = await useCases.withdrawalUseCase.execute(email: email)
+        
+        switch result {
+        case .success:
+            let isAccessDeleted = KeyChainManager.shared.deleteAccessToken()
+            let isRefreshDeleted = KeyChainManager.shared.deleteRefreshToken()
+            
+            appState.navigation.reset()
+            if isAccessDeleted && isRefreshDeleted {
+                appState.navigation.push(to: .root)
+            } else {
+                appState.navigation.push(to: .login)
+            }
+        case .failure(let error):
+            appState.system.showAlert = true
+            appState.system.errorMessage = error.localizedDescription()
+        }
+    }
+    
+    // MARK: - Helper Functions
+    
+    /// 키보드를 숨깁니다.
     func hideKeyboard() {
         UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
     }
     
-    /// tag id --> tag 맵핑하는 함수
-    private func getTags(from tagIDs: [Int]) -> [Tag] {
-        return tags.filter { tagIDs.contains($0.id) }
-    }
-    
+    /// 메인 관련 데이터를 초기화합니다.
     func clearMain() {
         memos = []
         tags = []
-        mainCurrentPage = 0
-        mainTotalPages = 1
         
         editorState = .create
         editorContent = ""
         editorTags = []
     }
     
+    /// 검색 관련 데이터를 초기화합니다.
     func clearSearch() {
         searchBarText = ""
         searchBarSelectedTags = []
         searchedMemos = []
         searchedTags = []
-        searchCurrentPage = 0
-        searchTotalPages = 1
+    }
+    
+    func mapTags(from tagIds: [UUID]) -> [Tag] {
+        return tags.filter { tagIds.contains($0.id) }
+    }
+    
+    func getSortedMemos() -> [Memo] {
+        switch sortMemo {
+        case .byCreate:
+            return memos.sorted { (memo1: Memo, memo2: Memo) -> Bool in
+                return memo1.createdAt < memo2.createdAt
+            }
+        case .byUpdate:
+            return memos.sorted { (memo1: Memo, memo2: Memo) -> Bool in
+                return memo1.updatedAt < memo2.updatedAt
+            }
+        }
+    }
+    
+    func getSortedRecommendedMemos() -> [Memo] {
+        switch sortMemo {
+        case .byCreate:
+            return recommendingMemos.sorted { (memo1: Memo, memo2: Memo) -> Bool in
+                return memo1.createdAt > memo2.createdAt
+            }
+        case .byUpdate:
+            return recommendingMemos.sorted { (memo1: Memo, memo2: Memo) -> Bool in
+                return memo1.updatedAt > memo2.updatedAt
+            }
+        }
+    }
+    
+    // MARK: - Helper Functions With AI
+    
+    /// 텍스트로부터 임베딩 벡터를 생성합니다.
+    func createEmbeddingVectorWithAI(text: String) async throws -> [Float] {
+        let embeddings = try AIModel.shared.encode(texts: [text])
+        return embeddings.first ?? []
+    }
+    
+    /// 메모 임베딩 벡터를 사용하여 태그 임베딩 벡터를 업데이트합니다.
+    func updateTagEmbeddingWithAI(memoEmbeddingVector: [Float], tagEmbeddingVector: [Float]) -> [Float] {
+        // 태그 임베딩을 메모 임베딩 쪽으로 1/100만큼 이동
+        let alpha: Float = 0.01
+        let updatedEmbedding = zip(tagEmbeddingVector, memoEmbeddingVector).map { $0 + alpha * $1 }
+        return updatedEmbedding
+    }
+    
+    /// 특정 메모에 속하는 모든 태그의 임베딩 벡터를 업데이트합니다.
+    /// CreateMemo할 때만 호출한다.
+    private func updateAssociatedTagsEmbedding(for memo: Memo) async {
+        for tagId in memo.tagIds {
+            if let currentTag = tags.first(where: { $0.id == tagId }) {
+                let tagEmbedding = currentTag.embeddingVector
+                
+                // 태그 임베딩 업데이트
+                let updatedEmbedding = updateTagEmbeddingWithAI(memoEmbeddingVector: memo.embeddingVector, tagEmbeddingVector: tagEmbedding)
+                
+                let currentDate = Date()
+                
+                // 태그 업데이트 요청
+                let result = await useCases.updateTagUseCase.execute(id: currentTag.id, name: currentTag.name, color: currentTag.color, embeddingVector: updatedEmbedding, createdAt: currentTag.createdAt, updatedAt: currentDate)
+                
+                switch result {
+                case .success(let updatedTag):
+                    var newTag = updatedTag
+                    newTag.embeddingVector = updatedEmbedding // 임베딩 벡터 할당
+                    
+                    // Main과 Search의 태그 목록에서 업데이트
+                    if let index = self.tags.firstIndex(where: { $0.id == updatedTag.id }) {
+                        self.tags[index] = newTag
+                    }
+                    if let index = self.searchedTags.firstIndex(where: { $0.id == updatedTag.id }) {
+                        self.searchedTags[index] = newTag
+                    }
+                    
+                case .failure(let error):
+                    print("태그 임베딩 업데이트 실패: \(error.localizedDescription)")
+                }
+            }
+        }
+    }
+    
+    // MARK: - Recommend Functions
+    
+    /// contentEmbeddingVector와 self.memos의 각 embeddingVector와 similarity를 구해서,
+    /// threshold 이상인 메모들을 similarity 기준 내림차순으로 정렬하여 반환한다.
+    func recommendMemosWithAI(contentEmbeddingVector: [Float]) -> [Memo] {
+        let threshold: Float = 0.5
+        let memoSimilarities = memos.compactMap { memo -> (memo: Memo, similarity: Float)? in
+            if let similarity = try? AIModel.shared.cosineSimilarity(vectorA: contentEmbeddingVector, vectorB: memo.embeddingVector),
+               similarity >= threshold {
+                return (memo, similarity)
+            } else {
+                return nil
+            }
+        }
+        let sortedMemos = memoSimilarities.sorted { $0.similarity > $1.similarity }
+        return sortedMemos.map { $0.memo }
+    }
+    
+    /// contentEmbeddingVector와 self.tags의 각 embeddingVector와 similarity를 구해서,
+    /// similarity 기준 내림차순으로 정렬된 태그 배열을 반환한다.
+    func recommendTagsWithAI(contentEmbeddingVector: [Float]) -> [Tag] {
+        let tagNames = tags.map { $0.name }
+        
+        guard let tagEmbeddings = try? AIModel.shared.encode(texts: tagNames) else { return [] }
+        
+        let tagSimilarities = zip(tags, tagEmbeddings).compactMap { (tag, embedding) -> (tag: Tag, similarity: Float)? in
+            if let similarity = try? AIModel.shared.cosineSimilarity(vectorA: contentEmbeddingVector, vectorB: embedding) {
+                return (tag, similarity)
+            } else {
+                return nil
+            }
+        }
+        
+        let sortedTags = tagSimilarities.sorted { $0.similarity > $1.similarity }
+        return sortedTags.map { $0.tag }
+    }
+    
+    /// editorContent와 editorTags를 읽고 recommendingMemos와 recommendingTags를 업데이트한다.
+    func recommendMemosAndTags() {
+        print("recommendMemosAndTags()")
+        // 편집창에 내용과 태그 모두 비어있으면 메모는 아무것도 추천 안하고, 태그는 관련성 없이 전부 추천함
+        if self.editorContent.isEmpty && self.editorTags.isEmpty {
+            self.recommendingMemos = []
+            self.recommendingTags = self.tags
+            return
+        }
+      
+        Task {
+            do {
+                // editorContent를 읽어서 임베딩 벡터 생성
+                let embeddingVector = try await createEmbeddingVectorWithAI(text: self.editorContent)
+                
+                // 편집창에 있는 태그의 ID들을 추출
+                let editorTagIds = editorTags.map { $0.id }
+                
+                // 그룹 A: 태그 매칭되는 메모 (update 상태 제외)
+                let groupAMemos = memos.filter { memo in
+                    memo.tagIds.contains(where: { editorTagIds.contains($0) }) &&
+                    !(editorState == .update(target: memo))
+                }
+                
+                // 그룹 B: AI 기반 추천 메모 중, 태그 매칭되지 않는 메모들 (이미 그룹 A에 포함된 것은 제외)
+                let groupBMemos = recommendMemosWithAI(contentEmbeddingVector: embeddingVector).filter { memo in
+                    !memo.tagIds.contains(where: { editorTagIds.contains($0) })
+                }
+                
+                // 최종 추천 메모: 그룹 A(태그 매칭) + 그룹 B(유사도 순)
+                let finalRecommendedMemos = groupAMemos + groupBMemos
+                
+                // 태그 추천
+                let recommendedTags = recommendTagsWithAI(contentEmbeddingVector: embeddingVector)
+                
+                // UI 업데이트
+                DispatchQueue.main.async {
+                    self.recommendingMemos = finalRecommendedMemos
+                    self.recommendingTags = recommendedTags
+                }
+            } catch {
+                appState.system.showAlert = true
+                appState.system.errorMessage = "추천 시스템에서 오류가 발생했습니다: \(error.localizedDescription)"
+            }
+        }
+    }
+
+
+    // MARK: - Search Functions
+    
+    /// searchTextEmbeddingVector와 self.memos의 각 embeddingVector와 similarity를 구해서,
+    /// threshold 이상인 메모들을 similarity 기준 내림차순으로 정렬하여 반환한다.
+    func searchMemosWithAI(searchTextEmbeddingVector: [Float]) -> [Memo] {
+        let threshold: Float = 0.5
+        let memoSimilarities = memos.compactMap { memo -> (memo: Memo, similarity: Float)? in
+            if let similarity = try? AIModel.shared.cosineSimilarity(vectorA: searchTextEmbeddingVector, vectorB: memo.embeddingVector),
+               similarity >= threshold {
+                return (memo, similarity)
+            } else {
+                return nil
+            }
+        }
+        let sortedMemos = memoSimilarities.sorted { $0.similarity > $1.similarity }
+        return sortedMemos.map { $0.memo }
+    }
+    
+    /// searchTextEmbeddingVector와 self.tags의 각 embeddingVector와 similarity를 구해서,
+    /// threshold 이상인 태그들을 similarity 기준 내림차순으로 정렬하여 반환한다.
+    func searchTagsWithAI(searchTextEmbeddingVector: [Float]) -> [Tag] {
+        let threshold: Float = 0.5
+        let tagSimilarities = tags.compactMap { tag -> (tag: Tag, similarity: Float)? in
+            if let similarity = try? AIModel.shared.cosineSimilarity(vectorA: searchTextEmbeddingVector, vectorB: tag.embeddingVector),
+               similarity >= threshold {
+                return (tag, similarity)
+            } else {
+                return nil
+            }
+        }
+        let sortedTags = tagSimilarities.sorted { $0.similarity > $1.similarity }
+        return sortedTags.map { $0.tag }
+    }
+    
+    /// searchBarText를 읽어서 searchedMemos와 searchedTags를 업데이트한다.
+    func searchMemosAndTags() {
+        // 검색창이 비어있으면 검색 결과를 전부 비우기
+        if self.searchBarText.isEmpty && self.searchBarSelectedTags.isEmpty {
+            self.searchedMemos = []
+            self.searchedTags = []
+            return
+        }
+        
+        Task {
+            do {
+                // searchBarText를 읽어서 임베딩 벡터 생성
+                let embeddingVector = try await createEmbeddingVectorWithAI(text: self.searchBarText)
+                
+                // 선택된 태그의 ID들을 추출
+                let selectedTagIds = searchBarSelectedTags.map { $0.id }
+                
+                // 그룹 A: 단순 문자열 매칭으로 검색 (선택된 태그를 모두 포함하고, content에 검색어가 포함된 메모)
+                let groupAMemos = memos.filter { memo in
+                    // 메모의 내용이 검색어를 포함하고
+                    (self.searchBarText.isEmpty || memo.content.contains(self.searchBarText)) &&
+                    // 선택된 태그를 모두 포함하는 경우
+                    selectedTagIds.allSatisfy { memo.tagIds.contains($0) }
+                }
+                
+                // 그룹 B: threshold 이상인 메모들을 유사도 기준 내림차순으로 정렬 (이미 정렬된 결과)
+                let groupBMemos = searchMemosWithAI(searchTextEmbeddingVector: embeddingVector)
+                
+                
+                // 그룹 A와 그룹 B를 중복 없이 결합 (그룹 A가 앞쪽에 오도록)
+                var combinedMemos = groupAMemos
+                for memo in groupBMemos where !groupAMemos.contains(where: { $0.id == memo.id }) {
+                    combinedMemos.append(memo)
+                }
+                
+                // 그룹 A: 이름 문자열 매칭으로 검색 (모든 태그 중 검색어 포함)
+                let groupATags = tags.filter { tag in
+                    tag.name.contains(self.searchBarText)
+                }
+                
+                // 그룹 B: 기존 유사도 기반 검색 결과 (threshold 이상인 태그들을 similarity 기준 내림차순)
+                let groupBTags = searchTagsWithAI(searchTextEmbeddingVector: embeddingVector)
+                
+                // 그룹 A와 그룹 B를 중복 없이 결합 (그룹 A가 앞쪽에 오도록)
+                var combinedTags = groupATags
+                for tag in groupBTags where !combinedTags.contains(where: { $0.id == tag.id }) {
+                    combinedTags.append(tag)
+                }
+                
+                // UI 업데이트
+                DispatchQueue.main.async {
+                    self.searchedMemos = combinedMemos
+                    self.searchedTags = combinedTags
+                }
+            } catch {
+                appState.system.showAlert = true
+                appState.system.errorMessage = "검색 시스템에서 오류가 발생했습니다: \(error.localizedDescription)"
+            }
+        }
     }
 }
